chore(suite): reorganize coin control file structure to resemble component structure
