+++ packages/suite/src/views/wallet/send/Options/BitcoinOptions/CoinControl/UtxoSelectionList/UtxoSelectionList.tsx
@@ -3,7 +3,7 @@ import { ReactNode } from 'react';
 import styled from 'styled-components';
 import { transparentize } from 'polished';
 
-import { selectAccountTransactionsWithNulls } from '@suite-common/wallet-core';
+import { selectAccountTransactions } from '@suite-common/wallet-core';
 import { Icon, variables, IconName } from '@trezor/components';
 import type { AccountUtxo } from '@trezor/connect';
 import { CSSColor } from '@trezor/theme';
@@ -11,7 +11,7 @@ import { CSSColor } from '@trezor/theme';
 import { useSelector } from 'src/hooks/suite';
 import { useSendFormContext } from 'src/hooks/wallet';
 
-import { UtxoSelection } from './UtxoSelection';
+import { UtxoSelection } from './UtxoSelection/UtxoSelection';
 
 const Wrapper = styled.section`
     border-bottom: 1px solid ${({ theme }) => theme.legacy.STROKE_GREY};
@@ -65,9 +65,7 @@ export const UtxoSelectionList = ({
 }: UtxoSelectionListProps) => {
     const { account } = useSendFormContext();
 
-    const accountTransactions = useSelector(state =>
-        selectAccountTransactionsWithNulls(state, account.key),
-    );
+    const accountTransactions = useSelector(state => selectAccountTransactions(state, account.key));
 
     return (
         <Wrapper>
@@ -89,7 +87,7 @@ export const UtxoSelectionList = ({
                 <UtxoSelection
                     key={`${utxo.txid}-${utxo.vout}`}
                     transaction={accountTransactions.find(
-                        transaction => transaction?.txid === utxo.txid,
+                        transaction => transaction.txid === utxo.txid,
                     )}
                     utxo={utxo}
                 />
